+++ R/plot-methods.R
@@ -31,14 +31,9 @@ plot.dsge_irf <- function(x, impulse = NULL, response = NULL,
   n_imp <- length(impulses)
   n_resp <- length(responses)
 
-  # Set up multi-panel plot
-  old_par <- par(no.readonly = TRUE)
-  on.exit(par(old_par))
-
-  par(mfrow = c(n_imp, n_resp),
-      mar = c(3, 3, 2, 1),
-      mgp = c(1.8, 0.5, 0),
-      cex = 0.7)
+  old_par <- graphics::par(no.readonly = TRUE)
+  on.exit(graphics::par(old_par))
+  .dsge_par_grid(n_imp, n_resp)
 
   has_ci <- "lower" %in% names(dat) && !all(is.na(dat$lower))
 
@@ -52,20 +47,20 @@ plot.dsge_irf <- function(x, impulse = NULL, response = NULL,
         ylim <- range(c(ylim, sub$lower, sub$upper), na.rm = TRUE)
       }
 
-      plot(sub$period, sub$value, type = "l", lwd = 2,
-           xlab = "Period", ylab = "",
-           main = paste0(imp, " -> ", resp),
-           ylim = ylim, ...)
+      # Empty frame first so gridlines and CI band sit underneath the line
+      graphics::plot(sub$period, sub$value, type = "n",
+                     xlab = "Period", ylab = "Response",
+                     main = sprintf("%s -> %s", imp, resp),
+                     ylim = ylim, ...)
+      .dsge_grid()
+      .dsge_zero_line()
 
       if (ci && has_ci) {
-        polygon(c(sub$period, rev(sub$period)),
-                c(sub$lower, rev(sub$upper)),
-                col = rgb(0.5, 0.5, 0.5, 0.2),
-                border = NA)
-        lines(sub$period, sub$value, lwd = 2)
+        .dsge_band(sub$period, sub$lower, sub$upper)
       }
 
-      abline(h = 0, lty = 2, col = "gray50")
+      graphics::lines(sub$period, sub$value,
+                      col = .DSGE_INK_PRIMARY, lwd = 1.8)
     }
   }
 }
@@ -85,19 +80,27 @@ plot.dsge_forecast <- function(x, ...) {
   vars <- unique(x$forecasts$variable)
   n_vars <- length(vars)
 
-  old_par <- par(no.readonly = TRUE)
-  on.exit(par(old_par))
+  old_par <- graphics::par(no.readonly = TRUE)
+  on.exit(graphics::par(old_par))
+  .dsge_par_grid(n_vars, 1L)
 
-  par(mfrow = c(n_vars, 1),
-      mar = c(3, 3, 2, 1),
-      mgp = c(1.8, 0.5, 0))
+  has_ci <- all(c("lower", "upper") %in% names(x$forecasts)) &&
+            !all(is.na(x$forecasts$lower))
 
   for (v in vars) {
     sub <- x$forecasts[x$forecasts$variable == v, ]
     sub <- sub[order(sub$period), ]
 
-    plot(sub$period, sub$value, type = "l", lwd = 2,
-         xlab = "Period", ylab = v,
-         main = paste("Forecast:", v), ...)
+    ylim <- range(sub$value, na.rm = TRUE)
+    if (has_ci) ylim <- range(c(ylim, sub$lower, sub$upper), na.rm = TRUE)
+
+    graphics::plot(sub$period, sub$value, type = "n",
+                   xlab = "Period", ylab = v,
+                   main = paste("Forecast:", v),
+                   ylim = ylim, ...)
+    .dsge_grid()
+    if (has_ci) .dsge_band(sub$period, sub$lower, sub$upper)
+    graphics::lines(sub$period, sub$value,
+                    col = .DSGE_INK_PRIMARY, lwd = 1.8)
   }
 }